[BACKEND-11] refactor(handler): rename handlers package to handler
