+++ internal/app/dependencies.go
@@ -3,7 +3,7 @@ package app
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/pp-sem6-team/backend/internal/config"
-	"github.com/pp-sem6-team/backend/internal/handlers"
+	handlers "github.com/pp-sem6-team/backend/internal/handler"
 	"github.com/pp-sem6-team/backend/internal/integration/storage/minio"
 	"github.com/pp-sem6-team/backend/internal/middleware"
 	"github.com/pp-sem6-team/backend/internal/repository/postgres"
